setup for RQ linode and update volumes patters
