+++ packages/manager/src/features/Volumes/VolumeAttachmentDrawer.tsx
@@ -10,7 +10,7 @@ import Select, { Item } from 'src/components/EnhancedSelect';
 import Notice from 'src/components/Notice';
 import { resetEventsPolling } from 'src/eventsPolling';
 import LinodeSelect from 'src/features/linodes/LinodeSelect';
-import { useAllLinodeConfigsQuery } from 'src/queries/linodes';
+import { useAllLinodeConfigsQuery } from 'src/queries/linodes/linodes';
 import { useGrants, useProfile } from 'src/queries/profile';
 import { useAttachVolumeMutation } from 'src/queries/volumes';
 import getAPIErrorsFor from 'src/utilities/getAPIErrorFor';
