+++ packages/manager/src/features/linodes/LinodesDetail/LinodeRescue/RescueDescription.tsx
@@ -5,7 +5,7 @@ import Typography from 'src/components/core/Typography';
 import Link from 'src/components/Link';
 import Notice from 'src/components/Notice';
 import { lishLaunch } from 'src/features/Lish/lishUtils';
-import { useLinodeFirewalls } from 'src/queries/linodes';
+import { useLinodeFirewalls } from 'src/queries/linodes/firewalls';
 
 const rescueDescription = {
   text: `If you suspect that your primary filesystem is corrupt, use the Linode Manager to boot your Linode into Rescue Mode. This is a safe environment for performing many system recovery and disk management tasks.`,
